+++ src/main/java/com/example/dietitian_plus/domain/product/dto/UpdateProductRequestDto.java
@@ -13,20 +13,20 @@ public class UpdateProductRequestDto {
     @JsonProperty("product_name")
     private String productName;
 
-    private Float kcal;
+    private Double kcal;
 
-    private Float fats;
+    private Double fats;
 
-    private Float carbs;
+    private Double carbs;
 
-    private Float protein;
+    private Double protein;
 
-    private Float fiber;
+    private Double fiber;
 
     @JsonProperty("glycemic_index")
-    private Float glycemicIndex;
+    private Double glycemicIndex;
 
     @JsonProperty("glycemic_load")
-    private Float glycemicLoad;
+    private Double glycemicLoad;
 
 }
